--- kh-final-project/src/main/java/com/kh/finalproject/controller/JoinController.java
@@ -1,31 +1,0 @@
-package com.kh.finalproject.controller;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import com.kh.finalproject.repository.MemberDao;
-import com.kh.finalproject.vo.member.MemberVo;
-
-@Controller
-public class JoinController {
-
-	@RequestMapping("/join")
-	public String join() {
-		return "member/join";
-	}
-	
-	@Autowired
-	private MemberDao memberDao; 
- 
-	@PostMapping(value = "/memberInsert")
-	public String memberInsert(@ModelAttribute MemberVo memberVo) {
-	memberDao.memberInsert(memberVo);
-	return "member/joinSuccess";
-	
-	}
-	
-	
-}
-
