--- kh-final-project/src/main/java/com/kh/finalproject/entity/CertDto.java
@@ -1,19 +1,0 @@
-package com.kh.finalproject.entity;
-
-import java.sql.Date;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CertDto {
-
-	private int auth_no;
-	private int auth_key;
-	private Date auth_date;
-}
