+++ kh-final-project/src/main/webapp/WEB-INF/views/member/login.jsp
@@ -21,7 +21,6 @@
 
 	.loginBtn{
 		border-radius:5px;
-		border-color:rgb
 		background:red;
 		color:#fff;
 		border:none;
